@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <WarningLevel>5</WarningLevel>
   </PropertyGroup>
 
@@ -17,14 +17,13 @@
 
   <ItemGroup>
     <PackageReference Include="Argu" Version="6.1.1" />
-    <PackageReference Include="Destructurama.FSharp" Version="1.2.0" />
-    <PackageReference Include="Equinox.CosmosStore" Version="3.0.7" />
-    <PackageReference Include="Equinox.EventStore" Version="3.0.7" />
-    <PackageReference Include="Equinox.MemoryStore" Version="3.0.7" />
-    <PackageReference Include="Equinox.Tools.TestHarness" Version="3.0.7" />
-    <PackageReference Include="FsCodec.SystemTextJson" Version="2.3.2" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.3" />
-    <PackageReference Include="Serilog.Sinks.Console" Version="4.0.0" />
+    <PackageReference Include="Equinox.CosmosStore" Version="4.0.0-rc.1" />
+    <PackageReference Include="Equinox.EventStoreDb" Version="4.0.0-rc.1" />
+    <PackageReference Include="Equinox.MemoryStore" Version="4.0.0-rc.1" />
+    <PackageReference Include="Equinox.Tools.TestHarness" Version="4.0.0-rc.1" />
+    <PackageReference Include="FsCodec.SystemTextJson" Version="3.0.0-rc.7.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.2" />
+    <PackageReference Include="Serilog.Sinks.Console" Version="4.0.1" />
     <PackageReference Include="Serilog.Sinks.Seq" Version="4.0.0" />
   </ItemGroup>
 
