@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <WarningLevel>5</WarningLevel>
   </PropertyGroup>
 
@@ -16,14 +16,13 @@
 
   <ItemGroup>
     <PackageReference Include="Argu" Version="6.1.1" />
-    <PackageReference Include="Destructurama.FSharp" Version="1.2.0" />
-    <PackageReference Include="Equinox.CosmosStore" Version="3.0.7" />
-    <PackageReference Include="Equinox.CosmosStore.Prometheus" Version="3.0.7" />
-    <PackageReference Include="FsCodec.SystemTextJson" Version="2.3.2" />
-    <PackageReference Include="Propulsion.CosmosStore" Version="2.12.2" />
-    <PackageReference Include="Propulsion.Feed" Version="2.12.2" />
+    <PackageReference Include="Equinox.CosmosStore" Version="4.0.0-rc.1" />
+    <PackageReference Include="Equinox.CosmosStore.Prometheus" Version="4.0.0-rc.1" />
+    <PackageReference Include="FsCodec.SystemTextJson" Version="3.0.0-rc.7.1" />
+    <PackageReference Include="Propulsion.CosmosStore" Version="3.0.0-beta.4" />
+    <PackageReference Include="Propulsion.Feed" Version="3.0.0-beta.4" />
     <PackageReference Include="Serilog.Sinks.Async" Version="1.5.0" />
-    <PackageReference Include="Serilog.Sinks.Console" Version="4.0.0" />
+    <PackageReference Include="Serilog.Sinks.Console" Version="4.0.1" />
   </ItemGroup>
 
 </Project>
