@@ -1,20 +1,22 @@
 module Patterns.Domain.Config
 
 let log = Serilog.Log.ForContext("isMetric", true)
-let createDecider stream = Equinox.Decider(log, stream, maxAttempts = 3)
+let resolveDecider cat = Equinox.Decider.resolve log cat
 
 module EventCodec =
 
     open FsCodec.SystemTextJson
 
     let private defaultOptions = Options.Create()
-    let create<'t when 't :> TypeShape.UnionContract.IUnionContract> () =
-        Codec.Create<'t>(options = defaultOptions).ToByteArrayCodec()
+    let gen<'t when 't :> TypeShape.UnionContract.IUnionContract> =
+        Codec.Create<'t>(options = defaultOptions)
+    let genJsonElement<'t when 't :> TypeShape.UnionContract.IUnionContract> =
+        CodecJsonElement.Create<'t>(options = defaultOptions)
 
 module Memory =
 
-    let create codec initial fold store =
-        Equinox.MemoryStore.MemoryStoreCategory(store, codec, fold, initial)
+    let create codec initial fold store : Equinox.Category<_, _, _> =
+        Equinox.MemoryStore.MemoryStoreCategory(store, FsCodec.Deflate.EncodeUncompressed codec, fold, initial)
 
 module Cosmos =
 
