@@ -1,7 +1,7 @@
 module Patterns.Domain.ListEpoch
 
 let [<Literal>] Category = "ListEpoch"
-let streamName = ListEpochId.toString >> FsCodec.StreamName.create Category
+let streamName id = struct (Category, ListEpochId.toString id)
 
 // NB - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
 [<RequireQualifiedAccess>]
@@ -12,7 +12,7 @@ module Events =
         | Closed
         | Snapshotted of        {| ids : ItemId[]; closed : bool |}
         interface TypeShape.UnionContract.IUnionContract
-    let codec = Config.EventCodec.create<Event>()
+    let codec, codecJe = Config.EventCodec.gen<Event>, Config.EventCodec.genJsonElement<Event>
 
 module Fold =
 
@@ -49,18 +49,16 @@ let decide shouldClose candidateIds = function
 // NOTE see feedSource for example of separating Service logic into Ingestion and Read Services in order to vary the folding and/or state held
 type Service internal
     (   shouldClose : ItemId[] -> ItemId[] -> bool, // let outer layers decide whether ingestion should trigger closing of the batch
-        resolve_ : Equinox.ResolveOption option -> ListEpochId -> Equinox.Decider<Events.Event, Fold.State>) =
-    let resolve = resolve_ None
-    let resolveStale = resolve_ (Some Equinox.AllowStale)
+        resolve : ListEpochId -> Equinox.Decider<Events.Event, Fold.State>) =
 
     /// Ingest the supplied items. Yields relevant elements of the post-state to enable generation of stats
     /// and facilitate deduplication of incoming items in order to avoid null store round-trips where possible
     member _.Ingest(epochId, items) : Async<ExactlyOnceIngester.IngestResult<_, _>> =
-        let decider = resolveStale epochId
-        /// NOTE decider which will initially transact against potentially stale cached state, which will trigger a
-        /// resync if another writer has gotten in before us. This is a conscious decision in this instance; the bulk
-        /// of writes are presumed to be coming from within this same process
-        decider.Transact(decide shouldClose items)
+        let decider = resolve epochId
+        // NOTE decider which will initially transact against potentially stale cached state, which will trigger a
+        // resync if another writer has gotten in before us. This is a conscious decision in this instance; the bulk
+        // of writes are presumed to be coming from within this same process
+        decider.Transact(decide shouldClose items, load = Equinox.AllowStale)
 
     /// Returns all the items currently held in the stream (Not using AllowStale on the assumption this needs to see updates from other apps)
     member _.Read epochId : Async<Fold.State> =
@@ -69,16 +67,9 @@ type Service internal
 
 module Config =
 
-    let private create_ shouldClose resolve = Service(shouldClose, resolve)
-    let private resolveStream opt = function
-        | Config.Store.Memory store ->
-            let cat = Config.Memory.create Events.codec Fold.initial Fold.fold store
-            fun sn -> cat.Resolve(sn, ?option = opt)
-        | Config.Store.Cosmos (context, cache) ->
-            let cat = Config.Cosmos.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
-            fun sn -> cat.Resolve(sn, ?option = opt)
-    let private resolveDecider store opt = streamName >> resolveStream opt store >> Config.createDecider
-    let private create__ shouldClose = resolveDecider >> create_ shouldClose
-    let create maxItemsPerEpoch =
+    let private (|Category|) = function
+        | Config.Store.Memory store -> Config.Memory.create Events.codec Fold.initial Fold.fold store
+        | Config.Store.Cosmos (context, cache) -> Config.Cosmos.createSnapshotted Events.codecJe Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
+    let create maxItemsPerEpoch (Category cat) =
         let shouldClose candidateItems currentItems = Array.length currentItems + Array.length candidateItems >= maxItemsPerEpoch
-        create__ shouldClose
+        Service(shouldClose, streamName >> Config.resolveDecider cat)
