@@ -1,7 +1,7 @@
 ﻿module TodoBackendTemplate.Aggregate
 
 let [<Literal>] Category = "Aggregate"
-let streamName (id: string) = FsCodec.StreamName.create Category id
+let streamName (id: string) = struct (Category, id)
 
 // NB - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
 module Events =
@@ -12,7 +12,7 @@ module Events =
         | Happened
         | Snapshotted of SnapshottedData
         interface TypeShape.UnionContract.IUnionContract
-    let codec = Config.EventCodec.create<Event>()
+    let codec, codecJe = Config.EventCodec.gen<Event>, Config.EventCodec.genJsonElement<Event>
 
 module Fold =
 
@@ -45,22 +45,22 @@ type Service internal (resolve : string -> Equinox.Decider<Events.Event, Fold.St
 
 module Config =
 
-    let private resolveStream = function
-#if (memoryStore || (!cosmos && !eventStore))
+    let private (|Category|) = function
+#if (memoryStore || (!cosmos && !dynamo && !eventStore))
         | Config.Store.Memory store ->
-            let cat = Config.Memory.create Events.codec Fold.initial Fold.fold store
-            cat.Resolve
+            Config.Memory.create Events.codec Fold.initial Fold.fold store
 #endif
 //#endif
 //#if cosmos
         | Config.Store.Cosmos (context, cache) ->
-            let cat = Config.Cosmos.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
-            cat.Resolve
+            Config.Cosmos.createSnapshotted Events.codecJe Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
+//#endif
+//#if dynamo
+        | Config.Store.Dynamo (context, cache) ->
+            Config.Dynamo.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
 //#endif
 //#if eventStore
         | Config.Store.Esdb (context, cache) ->
-            let cat = Config.Esdb.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
-            cat.Resolve
+            Config.Esdb.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
 //#endif
-    let private resolveDecider store = streamName >> resolveStream store >> Config.createDecider
-    let create = resolveDecider >> Service
+    let create (Category cat) = Service(streamName >> Config.resolveDecider cat)
