@@ -34,7 +34,7 @@ type SliceDto = { closed : bool; tickets : ItemDto[]; position : TicketsCheckpoi
 type Session(client: HttpClient) =
 
     member _.Send(req : HttpRequestMessage) : Async<HttpResponseMessage> =
-        client.Send(req)
+        client.Send2(req)
 
 type TicketsClient(session: Session) =
 
@@ -69,7 +69,7 @@ type TicketsFeed(baseUri) =
     let tickets = Session(client).Tickets
 
     // TODO add retries - consumer loop will abort if this throws
-    member _.Poll(trancheId, pos) : Async<Propulsion.Feed.Page<byte[]>> = async {
+    member _.Poll(trancheId, pos) : Async<Propulsion.Feed.Page<Propulsion.Streams.Default.EventBody>> = async {
         let checkpoint = TicketsCheckpoint.ofPosition pos
         let! pg = tickets.Poll(TrancheId.toFcId trancheId, checkpoint)
         let baseIndex = TicketsCheckpoint.toStreamIndex pg.position
