@@ -5,13 +5,11 @@ open Serilog
 open System
 
 exception MissingArg of message : string with override this.Message = this.message
+let missingArg msg = raise (MissingArg msg)
 
 type Configuration(tryGet) =
 
-    let get key =
-        match tryGet key with
-        | Some value -> value
-        | None -> raise (MissingArg (sprintf "Missing Argument/Environment Variable %s" key))
+    let get key = match tryGet key with Some value -> value | None -> missingArg $"Missing Argument/Environment Variable %s{key}"
 
     member _.CosmosConnection =             get "EQUINOX_COSMOS_CONNECTION"
     member _.CosmosDatabase =               get "EQUINOX_COSMOS_DATABASE"
@@ -32,7 +30,7 @@ module Args =
         | [<AltCommandLine "-k"; Unique>]   MaxKib of int
         | [<CliPrefix(CliPrefix.None); AltCommandLine "cosmos"; Unique(*ExactlyOnce is not supported*); Last>] SrcCosmos of ParseResults<CosmosSourceParameters>
         interface IArgParserTemplate with
-            member a.Usage = a |> function
+            member p.Usage = p |> function
                 | Verbose ->                "request Verbose Logging. Default: off"
                 | SyncVerbose ->            "request Logging for Sync operations (Writes). Default: off"
                 | PrometheusPort _ ->       "port from which to expose a Prometheus /metrics endpoint. Default: off"
@@ -42,20 +40,20 @@ module Args =
                 | RuThreshold _ ->          "minimum request charge required to log. Default: 0"
                 | MaxKib _ ->               "max KiB to submit to Sync operation. Default: 512"
                 | SrcCosmos _ ->            "Cosmos input parameters."
-    and Arguments(c : Configuration, a : ParseResults<Parameters>) =
-        member val Verbose =                a.Contains Parameters.Verbose
-        member val SyncLogging =            a.Contains SyncVerbose, a.TryGetResult RuThreshold
-        member val PrometheusPort =         a.TryGetResult PrometheusPort
-        member val ProcessorName =          a.GetResult ProcessorName
-        member val MaxReadAhead =           a.GetResult(MaxReadAhead, 32)
-        member val MaxWriters =             a.GetResult(MaxWriters, 4)
-        member val MaxBytes =               a.GetResult(MaxKib, 512) * 1024
+    and Arguments(c : Configuration, p : ParseResults<Parameters>) =
+        member val Verbose =                p.Contains Parameters.Verbose
+        member val SyncLogging =            p.Contains SyncVerbose, p.TryGetResult RuThreshold
+        member val PrometheusPort =         p.TryGetResult PrometheusPort
+        member val ProcessorName =          p.GetResult ProcessorName
+        member val MaxReadAhead =           p.GetResult(MaxReadAhead, 32)
+        member val MaxWriters =             p.GetResult(MaxWriters, 4)
+        member val MaxBytes =               p.GetResult(MaxKib, 512) * 1024
         member val StatsInterval =          TimeSpan.FromMinutes 1.
         member val StateInterval =          TimeSpan.FromMinutes 5.
         member val Source : CosmosSourceArguments =
-            match a.TryGetSubCommand() with
-            | Some (SrcCosmos cosmos) -> CosmosSourceArguments (c, cosmos)
-            | _ -> raise (MissingArg "Must specify cosmos for SrcCosmos")
+            match p.GetSubCommand() with
+            | SrcCosmos cosmos -> CosmosSourceArguments(c, cosmos)
+            | _ -> missingArg "Must specify cosmos for SrcCosmos"
         member x.DestinationArchive = x.Source.Archive
         member x.MonitoringParams() =
             let srcC = x.Source
@@ -64,7 +62,7 @@ module Args =
                 match srcC.LeaseContainer, dstC.LeaseContainerId with
                 | _, None ->        srcC.ConnectLeases()
                 | None, Some dc ->  dstC.ConnectLeases dc
-                | Some _, Some _ -> raise (MissingArg "LeaseContainerSource and LeaseContainerDestination are mutually exclusive - can only store in one database")
+                | Some _, Some _ -> missingArg "LeaseContainerSource and LeaseContainerDestination are mutually exclusive - can only store in one database"
             Log.Information("Archiving... {dop} writers, max {maxReadAhead} batches read ahead, max write batch {maxKib} KiB", x.MaxWriters, x.MaxReadAhead, x.MaxBytes / 1024)
             Log.Information("ChangeFeed {processorName} Leases Database {db} Container {container}. MaxItems limited to {maxItems}",
                 x.ProcessorName, leases.Database.Id, leases.Id, Option.toNullable srcC.MaxItems)
@@ -75,7 +73,7 @@ module Args =
     and [<NoEquality; NoComparison>] CosmosSourceParameters =
         | [<AltCommandLine "-V"; Unique>]   Verbose
         | [<AltCommandLine "-Z"; Unique>]   FromTail
-        | [<AltCommandLine "-mi"; Unique>]  MaxItems of int
+        | [<AltCommandLine "-b"; Unique>]   MaxItems of int
         | [<AltCommandLine "-l"; Unique>]   LagFreqM of float
         | [<AltCommandLine "-a"; Unique>]   LeaseContainer of string
 
@@ -89,7 +87,7 @@ module Args =
 
         | [<CliPrefix(CliPrefix.None); AltCommandLine "cosmos"; Unique(*ExactlyOnce is not supported*); Last>] DstCosmos of ParseResults<CosmosSinkParameters>
         interface IArgParserTemplate with
-            member a.Usage = a |> function
+            member p.Usage = p |> function
                 | Verbose ->                "request Verbose Change Feed Processor Logging. Default: off"
                 | FromTail ->               "(iff the Consumer Name is fresh) - force skip to present Position. Default: Never skip an event."
                 | MaxItems _ ->             "maximum item count to request from feed. Default: unlimited"
@@ -105,31 +103,31 @@ module Args =
                 | RetriesWaitTime _ ->      "specify max wait-time for retry when being throttled by Cosmos in seconds. Default: 30."
 
                 | DstCosmos _ ->            "CosmosDb Sink parameters."
-    and CosmosSourceArguments(c : Configuration, a : ParseResults<CosmosSourceParameters>) =
-        let discovery =                     a.TryGetResult CosmosSourceParameters.Connection |> Option.defaultWith (fun () -> c.CosmosConnection) |> Equinox.CosmosStore.Discovery.ConnectionString
-        let mode =                          a.TryGetResult CosmosSourceParameters.ConnectionMode
-        let timeout =                       a.GetResult(CosmosSourceParameters.Timeout, 5.) |> TimeSpan.FromSeconds
-        let retries =                       a.GetResult(CosmosSourceParameters.Retries, 5)
-        let maxRetryWaitTime =              a.GetResult(CosmosSourceParameters.RetriesWaitTime, 30.) |> TimeSpan.FromSeconds
+    and CosmosSourceArguments(c : Configuration, p : ParseResults<CosmosSourceParameters>) =
+        let discovery =                     p.TryGetResult CosmosSourceParameters.Connection |> Option.defaultWith (fun () -> c.CosmosConnection) |> Equinox.CosmosStore.Discovery.ConnectionString
+        let mode =                          p.TryGetResult CosmosSourceParameters.ConnectionMode
+        let timeout =                       p.GetResult(CosmosSourceParameters.Timeout, 5.) |> TimeSpan.FromSeconds
+        let retries =                       p.GetResult(CosmosSourceParameters.Retries, 5)
+        let maxRetryWaitTime =              p.GetResult(CosmosSourceParameters.RetriesWaitTime, 30.) |> TimeSpan.FromSeconds
         let connector =                     Equinox.CosmosStore.CosmosStoreConnector(discovery, timeout, retries, maxRetryWaitTime, ?mode = mode)
-        let database =                      a.TryGetResult CosmosSourceParameters.Database |> Option.defaultWith (fun () -> c.CosmosDatabase)
-        member val ContainerId =            a.GetResult CosmosSourceParameters.Container
+        let database =                      p.TryGetResult CosmosSourceParameters.Database |> Option.defaultWith (fun () -> c.CosmosDatabase)
+        member val ContainerId =            p.GetResult CosmosSourceParameters.Container
         member x.MonitoredContainer() =     connector.ConnectMonitored(database, x.ContainerId)
 
-        member val FromTail =               a.Contains CosmosSourceParameters.FromTail
-        member val MaxItems =               a.TryGetResult MaxItems
-        member val LagFrequency : TimeSpan = a.GetResult(LagFreqM, 1.) |> TimeSpan.FromMinutes
-        member val LeaseContainer =         a.TryGetResult CosmosSourceParameters.LeaseContainer
-        member val Verbose =                a.Contains Verbose
+        member val FromTail =               p.Contains CosmosSourceParameters.FromTail
+        member val MaxItems =               p.TryGetResult MaxItems
+        member val LagFrequency : TimeSpan = p.GetResult(LagFreqM, 1.) |> TimeSpan.FromMinutes
+        member val LeaseContainer =         p.TryGetResult CosmosSourceParameters.LeaseContainer
+        member val Verbose =                p.Contains Verbose
         member private _.ConnectLeases containerId = connector.CreateUninitialized(database, containerId)
         member x.ConnectLeases() =          match x.LeaseContainer with
                                             | None ->    x.ConnectLeases(x.ContainerId + "-aux")
                                             | Some sc -> x.ConnectLeases(sc)
 
         member val Archive =
-            match a.TryGetSubCommand() with
-            | Some (DstCosmos cosmos) -> CosmosSinkArguments (c, cosmos)
-            | _ -> raise (MissingArg "Must specify cosmos for Sink")
+            match p.GetSubCommand() with
+            | DstCosmos cosmos -> CosmosSinkArguments(c, cosmos)
+            | _ -> missingArg "Must specify cosmos for Sink"
     and [<NoEquality; NoComparison>] CosmosSinkParameters =
         | [<AltCommandLine "-m">]           ConnectionMode of Microsoft.Azure.Cosmos.ConnectionMode
         | [<AltCommandLine "-s">]           Connection of string
@@ -140,7 +138,7 @@ module Args =
         | [<AltCommandLine "-r">]           Retries of int
         | [<AltCommandLine "-rt">]          RetriesWaitTime of float
         interface IArgParserTemplate with
-            member a.Usage = a |> function
+            member p.Usage = p |> function
                 | ConnectionMode _ ->       "override the connection mode. Default: Direct."
                 | Connection _ ->           "specify a connection string for a Cosmos account. (optional if environment variable EQUINOX_COSMOS_CONNECTION specified)"
                 | Database _ ->             "specify a database name for Cosmos account. (optional if environment variable EQUINOX_COSMOS_DATABASE specified)"
@@ -149,18 +147,18 @@ module Args =
                 | Timeout _ ->              "specify operation timeout in seconds. Default: 5."
                 | Retries _ ->              "specify operation retries. Default: 0."
                 | RetriesWaitTime _ ->      "specify max wait-time for retry when being throttled by Cosmos in seconds. Default: 5."
-    and CosmosSinkArguments(c : Configuration, a : ParseResults<CosmosSinkParameters>) =
-        let discovery =                     a.TryGetResult Connection |> Option.defaultWith (fun () -> c.CosmosConnection) |> Equinox.CosmosStore.Discovery.ConnectionString
-        let mode =                          a.TryGetResult ConnectionMode
-        let timeout =                       a.GetResult(CosmosSinkParameters.Timeout, 5.) |> TimeSpan.FromSeconds
-        let retries =                       a.GetResult(CosmosSinkParameters.Retries, 0)
-        let maxRetryWaitTime =              a.GetResult(RetriesWaitTime, 5.) |> TimeSpan.FromSeconds
+    and CosmosSinkArguments(c : Configuration, p : ParseResults<CosmosSinkParameters>) =
+        let discovery =                     p.TryGetResult Connection |> Option.defaultWith (fun () -> c.CosmosConnection) |> Equinox.CosmosStore.Discovery.ConnectionString
+        let mode =                          p.TryGetResult ConnectionMode
+        let timeout =                       p.GetResult(CosmosSinkParameters.Timeout, 5.) |> TimeSpan.FromSeconds
+        let retries =                       p.GetResult(CosmosSinkParameters.Retries, 0)
+        let maxRetryWaitTime =              p.GetResult(RetriesWaitTime, 5.) |> TimeSpan.FromSeconds
         let connector =                     Equinox.CosmosStore.CosmosStoreConnector(discovery, timeout, retries, maxRetryWaitTime, ?mode = mode)
-        let database =                      a.TryGetResult Database |> Option.defaultWith (fun () -> c.CosmosDatabase)
-        let container =                     a.TryGetResult Container |> Option.defaultWith (fun () -> c.CosmosContainer)
+        let database =                      p.TryGetResult Database |> Option.defaultWith (fun () -> c.CosmosDatabase)
+        let container =                     p.TryGetResult Container |> Option.defaultWith (fun () -> c.CosmosContainer)
         member _.Connect() =                connector.ConnectStore("Destination", database, container)
 
-        member val LeaseContainerId =       a.TryGetResult LeaseContainer
+        member val LeaseContainerId =       p.TryGetResult LeaseContainer
         member _.ConnectLeases containerId = connector.CreateUninitialized(database, containerId)
 
     /// Parse the commandline; can throw exceptions in response to missing arguments and/or `-h`/`--help` args
@@ -183,11 +181,13 @@ let build (args : Args.Arguments, log) =
     let archiverSink =
         let context = args.DestinationArchive.Connect() |> Async.RunSynchronously |> CosmosStoreContext.create
         let eventsContext = Equinox.CosmosStore.Core.EventsContext(context, Config.log)
-        CosmosStoreSink.Start(log, args.MaxReadAhead, eventsContext, args.MaxWriters, args.StatsInterval, args.StateInterval, (*purgeInterval=TimeSpan.FromMinutes 10.,*) maxBytes = args.MaxBytes)
+        CosmosStoreSink.Start(log, args.MaxReadAhead, eventsContext, args.MaxWriters, args.StatsInterval, args.StateInterval,
+                              purgeInterval=TimeSpan.FromMinutes 10., maxBytes = args.MaxBytes)
     let source =
         let observer = CosmosStoreSource.CreateObserver(log, archiverSink.StartIngester, Seq.collect Handler.selectArchivable)
         let monitored, leases, processorName, startFromTail, maxItems, lagFrequency = args.MonitoringParams()
-        CosmosStoreSource.Start(log, monitored, leases, processorName, observer, startFromTail, ?maxItems=maxItems, lagReportFreq=lagFrequency)
+        CosmosStoreSource.Start(log, monitored, leases, processorName, observer,
+                                startFromTail = startFromTail, ?maxItems=maxItems, lagReportFreq=lagFrequency)
     archiverSink, source
 
 // A typical app will likely have health checks etc, implying the wireup would be via `endpoints.MapMetrics()` and thus not use this ugly code directly
@@ -197,25 +197,23 @@ let startMetricsServer port : IDisposable =
     Log.Information("Prometheus /metrics endpoint on port {port}", port)
     { new IDisposable with member x.Dispose() = ms.Stop(); (metricsServer :> IDisposable).Dispose() }
 
-open Propulsion.CosmosStore.Infrastructure // AwaitKeyboardInterruptAsTaskCancelledException
+open Propulsion.Internal // AwaitKeyboardInterruptAsTaskCanceledException
 
 let run (args : Args.Arguments) = async {
     let log = (Log.forGroup args.ProcessorName).ForContext<Propulsion.Streams.Scheduling.StreamSchedulingEngine>()
     let sink, source = build (args, log)
     use _metricsServer : IDisposable = args.PrometheusPort |> Option.map startMetricsServer |> Option.toObj
-    return! Async.Parallel [
-        Async.AwaitKeyboardInterruptAsTaskCancelledException()
-        source.AwaitWithStopOnCancellation()
-        sink.AwaitWithStopOnCancellation() ]
-    |> Async.Ignore<unit[]>
-}
+    return! [|  Async.AwaitKeyboardInterruptAsTaskCanceledException()
+                source.AwaitWithStopOnCancellation()
+                sink.AwaitWithStopOnCancellation()
+            |] |> Async.Parallel |> Async.Ignore<unit array> }
 
 [<EntryPoint>]
 let main argv =
     try let args = Args.parse EnvVar.tryGet argv
         try Log.Logger <- LoggerConfiguration().Configure(AppName, args.Verbose, args.SyncLogging).CreateLogger()
             try run args |> Async.RunSynchronously; 0
-            with e when not (e :? MissingArg) -> Log.Fatal(e, "Exiting"); 2
+            with e when not (e :? MissingArg) && not (e :? System.Threading.Tasks.TaskCanceledException) -> Log.Fatal(e, "Exiting"); 2
         finally Log.CloseAndFlush()
     with MissingArg msg -> eprintfn "%s" msg; 1
         | :? Argu.ArguParseException as e -> eprintfn "%s" e.Message; 1
