@@ -11,7 +11,7 @@ type Outcome =
     | NotApplicable of count : int
 
 /// Gathers stats based on the outcome of each Span processed for emission, at intervals controlled by `StreamsConsumer`
-type Stats(log, statsInterval, stateInterval) =
+type Stats(log, statsInterval, stateInterval, verboseStore, ?logExternalStats) =
     inherit Propulsion.Streams.Stats<Outcome>(log, statsInterval, stateInterval)
 
     let mutable ok, skipped, na = 0, 0, 0
@@ -21,24 +21,29 @@ type Stats(log, statsInterval, stateInterval) =
         | Outcome.Skipped count -> skipped <- skipped + count
         | Outcome.NotApplicable count -> na <- na + count
     override _.HandleExn(log, exn) =
-        log.Information(exn, "Unhandled")
+        Exception.dump verboseStore log exn
 
     override _.DumpStats() =
         base.DumpStats()
         if ok <> 0 || skipped <> 0 || na <> 0 then
             log.Information(" used {ok} skipped {skipped} n/a {na}", ok, skipped, na)
             ok <- 0; skipped <- 0; na <- 0
+        logExternalStats |> Option.iter (fun dumpTo -> dumpTo log)
 
 #if blank
-let handle (stream, span : Propulsion.Streams.StreamSpan<_>) = async {
+let categoryFilter = function
+    | sn when sn = "Todos" -> true
+    | _ -> false
+    
+let handle struct (stream, span : Propulsion.Streams.StreamSpan<_>) = async {
     match stream, span with
     | FsCodec.StreamName.CategoryAndId ("Todos", id), _ ->
         let ok = true
         // "TODO: add handler code"
         match ok with
-        | true -> return Propulsion.Streams.SpanResult.AllProcessed, Outcome.Ok (1, span.events.Length - 1)
-        | false -> return Propulsion.Streams.SpanResult.AllProcessed, Outcome.Skipped span.events.Length
-    | _ -> return Propulsion.Streams.AllProcessed, Outcome.NotApplicable span.events.Length }
+        | true -> return struct (Propulsion.Streams.SpanResult.AllProcessed, Outcome.Ok (1, span.Length - 1))
+        | false -> return Propulsion.Streams.SpanResult.AllProcessed, Outcome.Skipped span.Length
+    | _ -> return Propulsion.Streams.AllProcessed, Outcome.NotApplicable span.Length }
 #else
 // map from external contract to internal contract defined by the aggregate
 let toSummaryEventData ( x : Contract.SummaryInfo) : TodoSummary.Events.SummaryData =
@@ -46,15 +51,31 @@ let toSummaryEventData ( x : Contract.SummaryInfo) : TodoSummary.Events.SummaryD
         [| for x in x.items ->
             { id = x.id; order = x.order; title = x.title; completed = x.completed } |] }
 
+let categoryFilter = function
+    | Todo.Reactions.Category -> true
+    | _ -> false
+
 let handle
         (sourceService : Todo.Service)
         (summaryService : TodoSummary.Service)
-        (stream, span : Propulsion.Streams.StreamSpan<_>) = async {
+        struct (stream, span : Propulsion.Streams.StreamSpan<_>) = async {
     match stream, span with
     | Todo.Reactions.Parse (clientId, events) when events |> Seq.exists Todo.Reactions.impliesStateChange ->
         let! version', summary = sourceService.QueryWithVersion(clientId, Contract.ofState)
         match! summaryService.TryIngest(clientId, version', toSummaryEventData summary) with
-        | true -> return Propulsion.Streams.SpanResult.OverrideWritePosition version', Outcome.Ok (1, span.events.Length - 1)
-        | false -> return Propulsion.Streams.SpanResult.OverrideWritePosition version', Outcome.Skipped span.events.Length
-    | _ -> return Propulsion.Streams.SpanResult.AllProcessed, Outcome.NotApplicable span.events.Length }
+        | true -> return struct (Propulsion.Streams.SpanResult.OverrideWritePosition version', Outcome.Ok (1, span.Length - 1))
+        | false -> return Propulsion.Streams.SpanResult.OverrideWritePosition version', Outcome.Skipped span.Length
+    | _ -> return Propulsion.Streams.SpanResult.AllProcessed, Outcome.NotApplicable span.Length }
 #endif
+
+type Config private () =
+    
+    static member StartSink(log : Serilog.ILogger, stats,
+                            handle : struct (FsCodec.StreamName * Propulsion.Streams.Default.StreamSpan)
+                                     -> Async<struct (Propulsion.Streams.SpanResult * 'Outcome)>,
+                            maxReadAhead : int, maxConcurrentStreams : int, ?wakeForResults, ?idleDelay, ?purgeInterval) =
+        Propulsion.Streams.Default.Config.Start(log, maxReadAhead, maxConcurrentStreams, handle, stats, stats.StatsInterval.Period,
+                                                ?wakeForResults = wakeForResults, ?idleDelay = idleDelay, ?purgeInterval = purgeInterval)
+    
+    static member StartSource(log, sink, sourceConfig) =
+        SourceConfig.start (log, Config.log) sink categoryFilter sourceConfig
