@@ -1,7 +1,7 @@
 ﻿module ConsumerTemplate.SkuSummary
 
 let [<Literal>] Category = "SkuSummary"
-let streamName (id : SkuId) = FsCodec.StreamName.create Category (SkuId.toString id)
+let streamName (id : SkuId) = struct (Category, SkuId.toString id)
 
 // NB - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
 module Events =
@@ -16,7 +16,7 @@ module Events =
         | Ingested of ItemData
         | Snapshotted of ItemData[]
         interface TypeShape.UnionContract.IUnionContract
-    let codec = EventCodec.create<Event>()
+    let codec = EventCodec.gen<Event>
 
 module Fold =
 
@@ -61,9 +61,6 @@ type Service internal (resolve : SkuId -> Equinox.Decider<Events.Event, Fold.Sta
 
 module Config =
 
-    let private resolveStream = function
-        | Config.Store.Cosmos (context, cache) ->
-            let cat = Config.Cosmos.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
-            cat.Resolve
-    let private resolveDecider store = streamName >> resolveStream store >> Config.createDecider
-    let create = resolveDecider >> Service
+    let private (|Category|) = function
+        | Config.Store.Cosmos (context, cache) -> Config.Cosmos.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
+    let create (Category cat) = Service(streamName >> Config.createDecider cat)
